requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
builtwith==1.3.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docopt==0.6.2
future==0.17.1
idna==2.8
lxml==4.4.1
pipreqs==0.4.9
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
yarg==0.1.9
